Ruby 1.9 compatability
Module.instance_methods returns an array of symbols in Ruby 1.9, as opposed to 1.8 which returns an array of strings. This was causing all of the tests related to instance methods to fail on 1.9.

Speaking of tests, 1.9 doesn't like relative paths, so I've fixed the require statement at the top of the test file.